requirements.txt
Flask==2.3.3
email-validator==2.1.0
dnspython==2.4.2
requests==2.31.0
Werkzeug==2.3.7